requirements.txt
Flask~=2.0.2
Werkzeug~=2.0.2
PyYAML~=6.0
DateTime~=4.3